requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==0.9.4
django-extensions==1.1.1
ipdb==0.7
ipython==1.0.0
